Add entries namming for some offloading targets.
This patch aims at making the compiler design more robust and immune to
the entries order requirement by making the linker script used by the
host linker to append to the target image the pointer to the host
entries as well as the target triple. The runtime can optionally look-up
this information after the target image, and build the target entries
names by appending the name of the host entry to the target name. In
order for this to work, each host entry contains a unique address to a
string with a unique string that identifies the entry (before the patch
it contains NULL ). Note that the map between host and target is still
made using the addresses which is still unique for each dynamic library.
The entries name is only used to look-up symbols in the same dynamic
library.